+++ backend/chainlit/teams/__init__.py
@@ -1,0 +1,6 @@
+try:
+    import botbuilder
+except ModuleNotFoundError:
+    raise ValueError(
+        "The botbuilder-core package is required to integrate Chainlit with a Slack app. Run `pip install botbuilder-core --upgrade`"
+    )