+++ frontend/src/components/molecules/auth/ProviderButton.tsx
@@ -19,6 +19,7 @@ function capitalizeFirstLetter(string: string) {
 function getProviderName(provider: string) {
   switch (provider) {
     case 'azure-ad':
+    case 'azure-ad-hybrid':
       return 'Microsoft';
     case 'github':
       return 'GitHub';
@@ -40,6 +41,7 @@ function renderProviderIcon(provider: string) {
     case 'github':
       return <GitHub />;
     case 'azure-ad':
+    case 'azure-ad-hybrid':
       return <Microsoft />;
     case 'okta':
       return <Okta />;